+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleDataprocInterpreter.scala
@@ -126,6 +126,11 @@ private[google2] class GoogleDataprocInterpreter[F[_]: StructuredLogger: Timer: 
   ): F[Option[Cluster]] = {
     val fa =
       F.delay(clusterControllerClient.getCluster(project.value, region.value, clusterName.value))
+        .map(Option(_))
+        .handleErrorWith {
+          case _: com.google.api.gax.rpc.NotFoundException => F.pure(none[Cluster])
+          case e                                           => F.raiseError[Option[Cluster]](e)
+        }
 
     ev.ask
       .flatMap { traceId =>
@@ -133,13 +138,8 @@ private[google2] class GoogleDataprocInterpreter[F[_]: StructuredLogger: Timer: 
           fa,
           Some(traceId),
           s"com.google.cloud.dataproc.v1.ClusterControllerClient.getCluster(${project.value}, ${region.value}, ${clusterName.value})",
-          Show.show[Cluster](c => s"${c.getStatus.toString}")
+          Show.show[Option[Cluster]](c => s"${c.map(_.getStatus.toString).getOrElse("Not found")}")
         )
-      }
-      .map(Option(_))
-      .handleErrorWith {
-        case _: com.google.api.gax.rpc.NotFoundException => F.pure(none[Cluster])
-        case e                                           => F.raiseError[Option[Cluster]](e)
       }
   }