+++ google2/CHANGELOG.md
@@ -15,13 +15,14 @@ Changed:
 - Expose `GoogleComputeService.fromCredential`
 - Added max retries to `SubscriberConfig`
 - Update `getCluster`, `getInstance`'s logging to cluster's status
+- Don't log as error when `getCluster`, `getInstance` returns NotFound
 
 Added:
 - Add `detachDisk`
 - Add `streamUploadBlob`
 - Add `listPodStatus` to `KubernetesService`, returns statuses of all pods belonging to a k8s cluster
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.11-ea9bf21"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.11-TRAVIS-REPLACE-ME"`
 
 ## 0.10
 Changed: