+++ google2/CHANGELOG.md
@@ -24,7 +24,7 @@ Added:
 - Add `listPodStatus` to `KubernetesService`, returns statuses of all pods belonging to a k8s cluster
 - Add `getServiceExternalIp` to `KubernetesService`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.11-TRAVIS-REPLACE-ME"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.11-2e155f0"`
 
 ## 0.10
 Changed: