+++ google2/CHANGELOG.md
@@ -25,7 +25,7 @@ Added:
 - Add `getServiceExternalIp` to `KubernetesService`
 - Add more retry logic to `GKEService`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.11-TRAVIS-REPLACE-ME"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.11-997b116"`
 
 ## 0.10
 Changed: