+++ google2/CHANGELOG.md
@@ -17,6 +17,7 @@ Changed:
 - Update `getCluster`, `getInstance`'s logging to cluster's status
 - Don't log as error when `getCluster`, `getInstance` returns NotFound
 - Return `None` if `instance`, `cluster` or `disk` doesn't exist when trying to `deleteInstance`, `deleteCluster` or `deleteDisk`
+- Support 2.13
 
 Added:
 - Add `detachDisk`